Moved exceptions from Models directory to Exceptions directory
