+++ app/Exceptions/Article/ArticleWasNotModerated.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Models\Article\Exceptions;
+namespace App\Exceptions\Article;
 
 use DomainException;
