+++ app/Exceptions/Comment/CommentNotCommentable.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Models\Comment\Exception;
+namespace App\Exceptions\Comment;
 
 use DomainException;
