+++ app/Models/User/User.php
@@ -9,17 +9,17 @@ use App\Events\User\EmailChanging;
 use App\Events\User\PasswordChanged;
 use App\Events\User\PasswordReset;
 use App\Events\User\UserRegistered;
+use App\Exceptions\Article\ArticleNotPublished;
+use App\Exceptions\User\BookmarkAlreadyCreated;
+use App\Exceptions\User\BookmarkNotCreated;
+use App\Exceptions\User\InvalidVerificationToken;
+use App\Exceptions\User\PasswordResetNotRequested;
+use App\Exceptions\User\RegistrationAlreadyVerified;
+use App\Exceptions\User\VerificationNotRequested;
+use App\Exceptions\User\VerificationTokenExpired;
 use App\Models\Article\Article;
-use App\Models\Article\Exceptions\ArticleNotPublished;
 use App\Models\Comment\Comment;
 use App\Models\Topic\Topic;
-use App\Models\User\Exceptions\BookmarkAlreadyCreated;
-use App\Models\User\Exceptions\BookmarkNotCreated;
-use App\Models\User\Exceptions\InvalidVerificationToken;
-use App\Models\User\Exceptions\PasswordResetNotRequested;
-use App\Models\User\Exceptions\RegistrationAlreadyVerified;
-use App\Models\User\Exceptions\VerificationNotRequested;
-use App\Models\User\Exceptions\VerificationTokenExpired;
 use Carbon\CarbonImmutable;
 use Illuminate\Auth\Authenticatable;
 use Illuminate\Contracts\Auth\Access\Authorizable as AuthorizableInterface;
