+++ tests/Unit/Model/User/VerifyNewEmailTest.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace Tests\Unit\Model\User;
 
 use App\Events\User\EmailChanged;
-use App\Models\User\Exceptions\InvalidVerificationToken;
-use App\Models\User\Exceptions\VerificationNotRequested;
-use App\Models\User\Exceptions\VerificationTokenExpired;
+use App\Exceptions\User\InvalidVerificationToken;
+use App\Exceptions\User\VerificationNotRequested;
+use App\Exceptions\User\VerificationTokenExpired;
 use App\Models\User\User;
 use App\Models\User\VerifyToken;
 use Illuminate\Foundation\Testing\RefreshDatabase;
